+++ src/test/java/dk/kb/present/TestFiles.java
@@ -104,4 +104,7 @@ public class TestFiles {
     public static final String CUMULUS_RECORD_0c02aa10 = "xml/copyright_extraction/0c02aa10-b657-11e6-aedf-00505688346e.xml";
     public static final String CUMULUS_RECORD_226d41a0 = "xml/copyright_extraction/226d41a0-5a83-11e6-8b8d-0016357f605f.xml";
 
+    // dr_archive_supplementary_rights_metadata
+    public static final String PVICA7_DR_ARCHIVE_SUPPLEMENTARY_RIGHTS_METADATA_DS_RADIO_83191087 = "internal_test_files/preservica7/dr_archive_supplementary_rights_metadata/dsradio/83191087-69b3-4f46-ab64-f230d971def2.xml";
+    public static final String PVICA7_HOMEMADE_DR_ARCHIVE_SUPPLEMENTARY_RIGHTS_METADATA_DS_TV_TID_TV_METER = "internal_test_files/homemade/dr_archive_supplementary_rights_metadata/dstv/tidTvmeter.xml";
 }